Kernel: Refactor init code
